fixing & enhance

+++ src/main/java/com/maybank/integratorapp/component/system/messageprocessor/AccountInquiryMessageProcessor.java
@@ -53,7 +53,7 @@ public class AccountInquiryMessageProcessor {
     private final String ProcessName = "AccountInquiryProcess";
 
 
-    public String processMessage(String message,int loggerId) {
+    public String processMessage(String message,long loggerId) {
         String responseXml = "";
         logger.SetLogParent(loggerId);
